% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Configuration.R
\name{.createConsoleAppenderForSandboxLogging}
\alias{.createConsoleAppenderForSandboxLogging}
\title{Create Console Appender for Sandbox Logging}
\usage{
.createConsoleAppenderForSandboxLogging(
  layout = ParallelLogger::layoutParallel
)
}
\arguments{
\item{layout}{A layout function for the logger. Defaults to \code{ParallelLogger::layoutParallel}.}
}
\value{
An appended object for logging.
}
\description{
Creates a console appender for sandbox logging with a specified layout.
}
