% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Configuration.R
\name{fcr_setUpLogger}
\alias{fcr_setUpLogger}
\title{Set Up Logger}
\usage{
fcr_setUpLogger(logsFolder = "logs")
}
\arguments{
\item{logsFolder}{A string representing the folder name for the logs.}
}
\value{
A logger object.
}
\description{
Sets up a logger with console and file appender for logging.
}
