% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analysisSettings_CohortDiagnostics.R
\name{mod_analysisSettings_CohortDiagnostics_server}
\alias{mod_analysisSettings_CohortDiagnostics_server}
\title{CohortDiagnosticss Analysis Settings Server}
\usage{
mod_analysisSettings_CohortDiagnostics_server(id, r_connectionHandler)
}
\arguments{
\item{id}{A string representing the module's namespace.}

\item{r_connectionHandler}{A reactive object that provides access to the cohort data and connection information.}
}
\value{
A reactive expression that returns the analysis settings as a list.
}
\description{
Server module for handling the logic of the cohort overlaps analysis settings UI. This module updates the UI elements based on the selected cohorts and returns the analysis settings as a reactive expression.
}
