% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analysisSettings_CohortDiagnostics.R
\name{mod_analysisSettings_CohortDiagnostics_ui}
\alias{mod_analysisSettings_CohortDiagnostics_ui}
\title{CohortDiagnostics Analysis Settings UI}
\usage{
mod_analysisSettings_CohortDiagnostics_ui(id)
}
\arguments{
\item{id}{A string representing the module's namespace.}
}
\value{
A Shiny UI element that can be included in a Shiny app.
}
\description{
UI module for configuring the settings for CohortDiagnostics analysis. This module allows users to select cohorts and configure various analysis options.
}
