% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_resultsVisualisation_CohortDiagnostics.R
\name{mod_cohortDiagnosticsVisualization_server}
\alias{mod_cohortDiagnosticsVisualization_server}
\title{CohortDiagnostics Results Visualization Server Module}
\usage{
mod_cohortDiagnosticsVisualization_server(pathToResultsDatabase)
}
\arguments{
\item{pathToResultsDatabase}{The path to the results database}
}
\value{
A server module for visualizing CohortDiagnostics results
}
\description{
This function creates a server module for visualizing CohortDiagnostics results.
}
