% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_resultsVisualisation_CohortDiagnostics.R
\name{mod_cohortDiagnosticsVisualization_ui}
\alias{mod_cohortDiagnosticsVisualization_ui}
\title{CohortDiagnostics Results Visualization Module}
\usage{
mod_cohortDiagnosticsVisualization_ui(pathToResultsDatabase)
}
\arguments{
\item{pathToResultsDatabase}{The path to the results database}
}
\description{
This module provides UI and server components for visualizing CohortDiagnostics results.
}
