% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fct_covariatesSelector.R
\name{mod_fct_covariateSelector_ui}
\alias{mod_fct_covariateSelector_ui}
\title{Covariate Selector UI Module}
\usage{
mod_fct_covariateSelector_ui(
  inputId,
  label = NULL,
  analysisIdsToShow = NULL,
  analysisIdsSelected = NULL
)
}
\arguments{
\item{inputId}{The input ID for the UI component.}

\item{label}{The label for the UI component.}

\item{analysisIdsToShow}{A vector of analysis IDs to show in the selector. If NULL, all analysis IDs are shown.}

\item{analysisIdsSelected}{A vector of analysis IDs to be selected by default. If NULL, all analysis IDs are selected.}
}
\value{
A shiny UI component for selecting covariates.
}
\description{
This module creates a UI component for selecting covariates based on analysis IDs.
}
