% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fct_temporalRanges.R
\name{mod_temporalRanges_server}
\alias{mod_temporalRanges_server}
\title{Temporal Ranges Server Module}
\usage{
mod_temporalRanges_server(
  id,
  session,
  startRanges = list(c(-50000, 50000), c(50000, 0))
)
}
\arguments{
\item{id}{The module ID}

\item{session}{The Shiny session object}

\item{startRanges}{A list of initial ranges to display}
}
\value{
A reactive expression containing the list of temporal ranges
}
\description{
This function creates a server module for managing temporal ranges.
}
