% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fct_temporalRanges.R
\name{mod_temporalRanges_ui}
\alias{mod_temporalRanges_ui}
\title{Temporal Ranges Module}
\usage{
mod_temporalRanges_ui(id)
}
\arguments{
\item{id}{The module ID}
}
\description{
This module creates a UI and server components for managing temporal ranges.
}
