% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_Run.R
\name{run_app}
\alias{run_app}
\title{Launch the Shiny Application}
\usage{
run_app(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{shiny::shinyApp}.}
}
\value{
A Shiny application object.
}
\description{
This function launches a Shiny application for CO2 analysis, configured with the provided modules.
}
